requirements.txt
Pillow==11.1.0
torch==2.6.0
transformers==4.49.0
